Move Effect and EffectInstance to entity\effect namespace
