+++ src/pocketmine/entity/effect/HealthBoostEffect.php
@@ -23,8 +23,6 @@ declare(strict_types=1);
 
 namespace pocketmine\entity\effect;
 
-use pocketmine\entity\Effect;
-use pocketmine\entity\EffectInstance;
 use pocketmine\entity\Living;
 
 class HealthBoostEffect extends Effect{
