+++ src/pocketmine/event/entity/EntityDamageByEntityEvent.php
@@ -23,7 +23,7 @@ declare(strict_types=1);
 
 namespace pocketmine\event\entity;
 
-use pocketmine\entity\Effect;
+use pocketmine\entity\effect\Effect;
 use pocketmine\entity\Entity;
 use pocketmine\entity\Living;
