+++ src/pocketmine/item/GoldenAppleEnchanted.php
@@ -23,8 +23,8 @@ declare(strict_types=1);
 
 namespace pocketmine\item;
 
-use pocketmine\entity\Effect;
-use pocketmine\entity\EffectInstance;
+use pocketmine\entity\effect\Effect;
+use pocketmine\entity\effect\EffectInstance;
 
 class GoldenAppleEnchanted extends GoldenApple{
